// src/app/route-paths.enum.ts
export enum RoutePaths {
  Employees = 'employees',
  EditEmployee = 'edit-employee',
  AddEmployee = 'add-employee',
  PastEmployees = 'past-employees',
  Conferences = 'conferences',
  AddConference = 'conferences/add',
  EditConference = 'conferences/edit',
  EmergencyContacts = 'emergency-contacts',
  Projects = 'projects',
  EditProject = 'projects/:id',
  EditeamAllocation = 'team-allocation/:id',
  VacationDays = 'vacation-days',
  NonWorkingHolidays = 'non-working-holidays',
  Salaries = 'salaries',
  EditSalary = 'salaries/edit',
  SalaryBonuses = 'salary-bonuses',
  AddSalaryBonuses = 'salary-bonuses/add',
  EditSalaryBonuses = 'salary-bonuses/edit',
  SalaryBonusesHistory = 'salary-bonuses-history',
  ReferralBonus = 'referrals',
  AddReferralBonus = 'add-referral',
  EditReferralBonus = 'referrals/:id',
  EquipmentRequest = 'equipment-request',
  EquipmentCheckoutRequest = 'equipment-checkout-request',
  EquipmentCompanies = 'equipment-companies',
  AddEquipmentCompany = 'equipment-companies/add',
  EditEquipmentCompany = 'equipment-companies/edit',
  EquipmentModelCategories = 'equipment-model-categories',
  AddEquipmentModelCategory = 'equipment-model-categories/add',
  EditEquipmentModelCategory = 'equipment-model-categories/edit',
  EquipmentAssetModel = 'equipment-models',
  AddEquipmentAssetModel = 'equipment-models/add',
  EditEquipmentAssetModel = 'equipment-models/edit',
  EquipmentSuppliers = 'equipment-suppliers',
  AddEquipmentSupplier = 'equipment-suppliers/add',
  EditEquipmentSupplier = 'equipment-suppliers/edit',
  EquipmentLocations = 'equipment-locations',
  AddEquipmentLocation = 'equipment-locations/add',
  EditEquipmentLocation = 'equipment-locations/edit',
  AddEquipment = 'equipments/add',
  EditEquipment = 'equipments/edit',
  Equipments = 'equipments',
  Books = 'books',
  AddBook = 'books/add-book',
  EditBook = 'books/:id',
  BookCheckouts = 'books/book-checkouts',
  AddBookCheckout = 'books/book-checkouts/add-book-checkout',
  EditBookCheckout = 'books/book-checkout/:id',
  Login = 'login',
  Settings = 'settings',
  Profile = 'profile',
  EmployeesSkills = 'employees-skills',
  SalarySlips = 'salary-slips',
  AddEmployeeInSkills = 'employees-skills/add-employee',
  EditUserInSkills = 'edit-employee',
  GoogleAuthCallback = 'connect/google/redirect',
  EmployeeDetails = 'employee-details/:id',
  TeamAllocation = 'team-allocation',
  VacationForm = 'vacation-form',
  MyVacationDays = 'my-vacation-days',
  VacationCalendar = 'vacation-calendar',
}